requirements.txt
alpaca-py==0.43.2
annotated-types==0.7.0
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
msgpack==1.1.2
numpy==2.4.1
pandas==3.0.0
prometheus_client==0.24.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
sseclient-py==1.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
